test: split useMessageActions.test.ts by the actions it groups
